提交学习笔记3.19

+++ JDBCProject_Supermarket/src/org/javasm/supermarket/bean/Product.java
@@ -11,7 +11,7 @@ public class Product {
   private Integer productStore;
   private String productImage;
   private boolean productStatus;
-  private Integer productDiscount;
+  private Double productDiscount;
   private java.util.Date createTime;
   private java.util.Date updateTime;
